[IMP]: Change data type integer to float in function field which calculate date diff in  crm and project issue module.
bzr revid: rch_open_erp-20100422064017-8dl3jiwwmjca90bh

+++ addons/crm/crm_opportunity.py
@@ -62,9 +62,9 @@ class crm_opportunity(osv.osv):
         @param ids: List of Openday’s IDs
         @return: difference between current date and log date
         @param context: A standard dictionary for contextual values
-        """        
-        cal_obj = self.pool.get('resource.calendar') 
-        res_obj = self.pool.get('resource.resource')     
+        """
+        cal_obj = self.pool.get('resource.calendar')
+        res_obj = self.pool.get('resource.resource')
 
         res = {}
         for opportunity in self.browse(cr, uid, ids , context):
@@ -92,7 +92,7 @@ class crm_opportunity(osv.osv):
 
                     duration = float(ans.days)
                     if opportunity.section_id.resource_calendar_id:
-                        duration =  float(ans.days) * 24                        
+                        duration =  float(ans.days) * 24
                         new_dates = cal_obj.interval_get(cr,
                             uid,
                             opportunity.section_id.resource_calendar_id and opportunity.section_id.resource_calendar_id.id or False,
@@ -104,9 +104,9 @@ class crm_opportunity(osv.osv):
                         date_until = mx.DateTime.strptime(date_until, '%Y-%m-%d %H:%M:%S')
                         for in_time, out_time in new_dates:
                             if in_time.date not in no_days:
-                                no_days.append(in_time.date)                            
+                                no_days.append(in_time.date)
                             if out_time > date_until:
-                                break                            
+                                break
                         duration =  len(no_days)
                 res[opportunity.id][field] = abs(int(duration))
         return res
@@ -137,9 +137,9 @@ class crm_opportunity(osv.osv):
 
         'date_open': fields.datetime('Opened', readonly=True),
         'day_open': fields.function(_compute_day, string='Days to Open', \
-                                method=True, multi='day_open', type="integer", store=True),
+                                method=True, multi='day_open', type="float", store=True),
         'day_close': fields.function(_compute_day, string='Days to Close', \
-                                method=True, multi='day_close', type="integer", store=True),
+                                method=True, multi='day_close', type="float", store=True),
          }
 
     def onchange_stage_id(self, cr, uid, ids, stage_id, context={}):
